requirements.txt
flask
flask-sqlalchemy==2.0
flask-wtf
flask-babel
flask-mail
flask-script
flask-oauthlib
flask-socketio
flask-testing
flask-redis
flask-compress
flask-cache
flask-pymongo
flask-bcrypt
flask-marshmallow
sqlalchemy==1.0.8
psycopg2
pillow
oauth
redis
python-slugify
mysqlclient
marshmallow
marshmallow-sqlalchemy
werkzeug==0.10.4
